backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
sqlalchemy==2.0.30
aiosqlite==0.20.0
pydantic-settings==2.2.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
